requirements.txt
streamlit>=1.32.0
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0
mysql-connector-python>=8.3.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.18.0
matplotlib>=3.8.0
fastapi>=0.100.0
uvicorn>=0.22.0
shap>=0.44.0
python-multipart>=0.0.6
